Create CompoundModule.cpp
NOTE: I use the float primitive type to do the calculations in this application but it is not advisable to use this type because the impreciseness of float numbers can cause errors that result in incorrect monetary values. E.g. 14.234 + 18.673 gives a value of  32.907 but the %.2f numeric formatting (2 decimal points) prints it out as 32.91, which is incorrect when considering the fractional part of the monetary value!!!